demos/palm/python/docs-agent/pyproject.toml
[tool.poetry]
name = "docs-agent"
version = "0.2.0"
description = ""
authors = ["Docs Agent contributors"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
rich = "^13.3.5"
Markdown = "^3.4.3"
beautifulsoup4 = "^4.12.2"
protobuf = ">=3.20"
chromadb = "==0.4.13"
sentence-transformers = "^2.2.2"
ratelimit = "^2.2.1"
absl-py = "^1.4.0"
python-frontmatter = "^1.0.0"
flatdict = "^4.0.1"
google-generativeai = "^0.3.1"
grpcio = "^1.57.0"
grpcio-tools = "^1.57.0"
uuid = "^1.30"
pytz = ">=2020.1"

[tool.poetry.group.dev.dependencies]
ipython = "^8.13.2"
pylint = "^2.17.4"
black = "^23.3.0"

[tool.poetry.group.internal]
optional = true

[tool.poetry.group.internal.dependencies]
seqio = "^0.0.16"

[tool.poetry.group.chatbot.dependencies]
flask = "^2.3.2"
flask-cors = "^3.0.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
